+++ RELEASENOTES.md
@@ -1,3 +1,10 @@
+# 1.6
+- Einstellungen für conf/FileSystemConfig.txt hinzugefügt
+- Einstellungen für IPEDConfig.txt hinzugefügt
+- Einstellungen für AudioTranslation überarbeitet
+- Einstellungen für ImageClassification überarbeitet
+- Einstellungen für Kategorien hinzugefügt
+
 # 1.5
 - Fehlerhafte Tooltips entfernt
 - Funktion zur Graphengenerierung implementiert